requirements.txt
Pillow
pyyaml
requests
requests_oauthlib